Relocate model migrations
